Modified CMakeLists. It works with Apple and Linux
Former-commit-id: e02c8f2ed499b7858bef216235ea8a74be9d9b9b

+++ clinica/lib/WeightedPrototypes/CPP_code/Density.cpp
@@ -52,30 +52,30 @@ int main(int argc, char** argv)
 
 	// Time
 	struct timeval start, end;
-	double delta;	
+	double delta;
 	gettimeofday(&start, NULL);
 
-	
+
 	// Parameters to see how many threads there are
 	Eigen::setNbThreads(0);; // Otherwise Eigen use OpenMP
 
 	int nThreads, tid;
 	#pragma omp parallel private(tid)
 	{
-		tid = omp_get_thread_num();	
-			
+		tid = omp_get_thread_num();
+
 		if (tid == 0)
 		{
 			nThreads = omp_get_num_threads();
-			//printf("Total number of threads: %d\n", nThreads);  
+			//printf("Total number of threads: %d\n", nThreads);
 		}
 	}
-	
+
 	omp_set_num_threads(nThreads); // in teoria non serve
-	
+
 	// Reading Parameters
 	char* PointsFibersFile = argv[1];
-	char* PointsMedoidsFile = argv[2];	
+	char* PointsMedoidsFile = argv[2];
 	char* TauFile = argv[3];
 	char* SurfaceFile = argv[4];
 	double lambda = atof(argv[5]);
@@ -94,7 +94,7 @@ int main(int argc, char** argv)
 	MatrixXf PointsMedoids;	// [NMedoidsx3]
 	unsigned int NMedoids;
 	MatrixXf GammaB;
-	int PI = 3.14159265358979323846;
+	double PI = 3.14159265358979323846;
 	VectorXf DensityB;
 	VectorXf DensityM;
 	MatrixXf GammaM;
@@ -104,28 +104,28 @@ int main(int argc, char** argv)
 
 
 // Surface
-	
+
 	vtkSmartPointer<vtkPolyDataReader> reader = vtkSmartPointer<vtkPolyDataReader>::New();
 	reader->SetFileName(SurfaceFile);
 	reader->Update();
 	vtkSmartPointer<vtkPolyData> polyData = reader->GetOutput();
 
 	NumberPointsSurface = polyData->GetNumberOfPoints();
-	
+
 	PointCoordinates.setZero(NumberPointsSurface,3);
 	for (unsigned int i = 0; i < NumberPointsSurface; i++)
 	{
-		 double p[3];		 
-		 polyData->GetPoint(i, p);		
+		 double p[3];
+		 polyData->GetPoint(i, p);
 		 for (int dim = 0; dim < 3; dim++)
 		 	PointCoordinates(i, dim) = p[dim];
 	}
 
 	//int NumFaces = polyData->GetNumberOfCells();
 
-//cout << "PointCoordinates: \n" << PointCoordinates << "\n" << endl;	
+//cout << "PointCoordinates: \n" << PointCoordinates << "\n" << endl;
 
-// Coordinates Points Bundle	
+// Coordinates Points Bundle
 
 	ifstream PointsFibersStream;
 	PointsFibersStream.open(PointsFibersFile,fstream::in | fstream::binary);
@@ -153,7 +153,7 @@ int main(int argc, char** argv)
 
 //cout << "PointsFibers: \n" << PointsFibers << "\n" <<  endl;
 
-// Coordinates Points Medoids	
+// Coordinates Points Medoids
 
 	ifstream PointsMedoidsStream;
 	PointsMedoidsStream.open(PointsMedoidsFile,fstream::in | fstream::binary);
@@ -189,7 +189,7 @@ int main(int argc, char** argv)
 		cerr << "Error! Could not open Tau file!" << endl;
 		return -1;
 	}
-	
+
 	ExactTau.setZero(NMedoids);
 	TauStream.read((char *)&ExactTau(0), NMedoids*4);
 	TauStream.close();
@@ -212,36 +212,36 @@ int main(int argc, char** argv)
 	for(unsigned int dim=0; dim<3; dim++)
 	{
 		MatrixXf MV;
-		MV.setZero(NumberPointsSurface,NFibers);	
-		#pragma omp parallel for shared(MV,PointCoordinates)	
+		MV.setZero(NumberPointsSurface,NFibers);
+		#pragma omp parallel for shared(MV,PointCoordinates)
 		for (unsigned int i=0; i<NFibers; i++)
 		{
-			MV.col(i)=PointCoordinates.col(dim);						
+			MV.col(i)=PointCoordinates.col(dim);
 		}
 
 		#pragma omp parallel for shared(MV,PointsFibers)
 		for (unsigned int i=0; i<NumberPointsSurface; i++)
 		{
-			MV.row(i) = MV.row(i) - PointsFibers.col(dim).transpose();									
-		}	
+			MV.row(i) = MV.row(i) - PointsFibers.col(dim).transpose();
+		}
 
 		MV=MV.array().pow(2);
-		GammaB=GammaB+MV;		
+		GammaB=GammaB+MV;
 	}
-	
+
 	PointsFibers.resize(1,1);
-	
-	GammaB=-GammaB/(2*lambda*lambda);	
-	GammaB=GammaB.array().exp();   
-	GammaB=GammaB*( ( 1/sqrt(8*PI*PI*PI) ) * ( 1/(lambda*lambda*lambda) ) );	
+
+	GammaB=-GammaB/(2*lambda*lambda);
+	GammaB=GammaB.array().exp();
+	GammaB=GammaB*( ( 1/sqrt(8*PI*PI*PI) ) * ( 1/(lambda*lambda*lambda) ) );
 
 	DensityB.setZero(NumberPointsSurface);
-	DensityB=(GammaB.rowwise().sum())/NFibers;  
+	DensityB=(GammaB.rowwise().sum())/NFibers;
 
-	GammaB.resize(1,1);	
+	GammaB.resize(1,1);
 
 	// Gamma Medoids
-	
+
 	try{
 		GammaM.setZero(NumberPointsSurface,NMedoids);
 	}
@@ -249,33 +249,33 @@ int main(int argc, char** argv)
 		cerr << "Error to allocate memory! Too many medoids!" << endl;
 		return -1;
 	}
-	
+
 	for(unsigned int dim=0; dim<3; dim++)
 	{
 		MatrixXf MV;
-		MV.setZero(NumberPointsSurface,NMedoids);	
-		#pragma omp parallel for shared(MV,PointCoordinates)	
+		MV.setZero(NumberPointsSurface,NMedoids);
+		#pragma omp parallel for shared(MV,PointCoordinates)
 		for (unsigned int i=0; i<NMedoids; i++)
 		{
-			MV.col(i)=PointCoordinates.col(dim);						
+			MV.col(i)=PointCoordinates.col(dim);
 		}
 
 		#pragma omp parallel for shared(MV,PointsFibers)
 		for (unsigned int i=0; i<NumberPointsSurface; i++)
 		{
-			MV.row(i) = MV.row(i) - PointsMedoids.col(dim).transpose();									
-		}	
+			MV.row(i) = MV.row(i) - PointsMedoids.col(dim).transpose();
+		}
 
 		MV=MV.array().pow(2);
-		GammaM=GammaM+MV;		
+		GammaM=GammaM+MV;
 	}
 
 	PointCoordinates.resize(1,1);
 	PointsMedoids.resize(1,1);
-	
-	GammaM=-GammaM/(2*lambda*lambda);	
-	GammaM=GammaM.array().exp();   
-	GammaM=GammaM*( ( 1/sqrt(8*PI*PI*PI) ) * ( 1/(lambda*lambda*lambda) ) );	
+
+	GammaM=-GammaM/(2*lambda*lambda);
+	GammaM=GammaM.array().exp();
+	GammaM=GammaM*( ( 1/sqrt(8*PI*PI*PI) ) * ( 1/(lambda*lambda*lambda) ) );
 
 	#pragma omp parallel for shared(GammaM,ExactTau)
 	for (unsigned int i=0; i<NumberPointsSurface; i++)
@@ -285,13 +285,13 @@ int main(int argc, char** argv)
 	}
 
 	DensityM.setZero(NumberPointsSurface);
-	DensityM=(GammaM.rowwise().sum())/(ExactTau.sum());  
+	DensityM=(GammaM.rowwise().sum())/(ExactTau.sum());
 
 	GammaM.resize(1,1);
-	
+
 	cout << "Number fibers: " << NFibers << ", sum tau: " << ExactTau.sum() << endl;
 
-	ExactTau.resize(1);	
+	ExactTau.resize(1);
 
 	ofstream DensityFibersBin;
 	DensityFibersBin.open("DensityFibers.bin", fstream::out | fstream::binary);
@@ -299,7 +299,7 @@ int main(int argc, char** argv)
       		float fib = DensityB[j];
       		DensityFibersBin.write((char *)(&fib),sizeof(float));
     	}
-  	DensityFibersBin.close();	    
+  	DensityFibersBin.close();
 
 	ofstream DensityMedoidsBin;
 	DensityMedoidsBin.open("DensityMedoids.bin", fstream::out | fstream::binary);
@@ -312,11 +312,8 @@ int main(int argc, char** argv)
 // TIMER
 	gettimeofday(&end, NULL);
 	delta = double(end.tv_sec  - start.tv_sec) + double(end.tv_usec - start.tv_usec) / 1.e6;
-	printf ("It took %f seconds \n",delta);	
+	printf ("It took %f seconds \n",delta);
 
 	return 0;
 
 } // end main
-
-
-
